requirements.txt
numpy
pandas
scipy
plotly
streamlit
matplotlib
pydantic
pyyaml
pytest
